Menü funktioniert

+++ reservation/src/Menu.java
@@ -1,9 +1,13 @@
+import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Menu {
 
     private String input;
     private Scanner scan;
+
+    private ArrayList<Admin> admins;
+    private ArrayList<Benutzer> users;
 
     private Admin admin1;
     private Admin admin2;
@@ -15,12 +19,19 @@ public class Menu {
 
         scan = new Scanner(System.in);
 
-        // Benutzer anlegen
-        admin1 = new Admin("Rolf","Meier",2345);
-        admin2 = new Admin("Lara","Croft",1345);
+        admins = new ArrayList<>();
+        users = new ArrayList<>();
 
-        user1 = new Benutzer("Frank","Larm",123);
-        user2 = new Benutzer("Anna","Mock",223);
+        // Benutzer anlegen
+        admin1 = new Admin("Rolf","Meier","apfel2",2345);
+        admin2 = new Admin("Lara","Croft","aprikose3",1345);
+        admins.add(admin1);
+        admins.add(admin2);
+
+        user1 = new Benutzer("Frank","Larm","melone1",123);
+        user2 = new Benutzer("Anna","Mock","banane5",223);
+        users.add(user1);
+        users.add(user2);
 
     }
 
@@ -37,10 +48,13 @@ public class Menu {
             switch (input.toLowerCase()){
                 case "a":
                     adminMenu();
+                    break;
                 case "u":
                     userMenu();
+                    break;
                 case "exit":
                     System.out.println("Applikation beendet");
+                    break;
                 default:
                     System.out.println("Ungültige Eingabe");
             }
@@ -48,13 +62,67 @@ public class Menu {
         }while (!input.equalsIgnoreCase("exit"));
     }
 
-
+    // Returnes "logged in" Admin
     public Admin adminLogin(){
-        Admin admin = null;
-        System.out.println("Admin Login");
-        System.out.println("------------------");
 
-        //Todo: Admin auswählen
+        Admin admin = null;
+        String vorname = null;
+        boolean loginCorrect = false;
+
+        do{
+            System.out.println("Admin Login");
+            System.out.println("------------------");
+
+            //Vorname
+            boolean nameFound = false;
+            do {
+                System.out.println("Vorname eingeben");
+                input = scan.nextLine();
+
+                for (Admin a: admins) {
+
+                    if(a.getVorname().equalsIgnoreCase(input)){
+                        nameFound = true;
+                        System.out.println("Vorname gefunden");
+                        vorname = a.getVorname();
+                    }
+                }
+
+                if(!nameFound){
+                    System.out.println("Vorname unbekannt");
+                }
+
+            }while (!(nameFound || input.equalsIgnoreCase("back")));
+
+
+            if(!input.equalsIgnoreCase("back")){
+
+                //Passwort
+                boolean passwordCorrect = false;
+                do {
+                    System.out.println("Passwort eingeben");
+                    input = scan.nextLine();
+
+                    for (Admin a: admins) {
+
+                        if(a.getPasswort().equals(input) && a.getVorname().equals(vorname)){
+                            passwordCorrect = true;
+                            System.out.println("Passwort korrekt");
+                            admin = a;
+                            loginCorrect = true;
+                        }
+                    }
+
+                    if(!passwordCorrect){
+                        System.out.println("Passwort falsch");
+                    }
+
+                }while (!(passwordCorrect || input.equalsIgnoreCase("back")));
+            }
+
+
+
+        }while(!(loginCorrect || input.equalsIgnoreCase("back")) );
 
         return admin;
     }
@@ -64,33 +132,42 @@ public class Menu {
 
         Admin admin = adminLogin();
 
-        do {
-           System.out.println("Admin Menü");
-           System.out.println("------------------");
-           System.out.println("Film erstellen         -> f");
-           System.out.println("Vorstellung erstellen  -> v");
-           System.out.println("Film löschen           -> fl");
-           System.out.println("Vorstellung löschen    -> vl");
-           System.out.println("Zurück ins Hauptmenü   -> exit");
+        if(admin != null){
 
-            input = scan.nextLine();
+            do {
+                System.out.println("Admin Menü");
+                System.out.println("------------------");
+                System.out.println("Film erstellen         -> f");
+                System.out.println("Vorstellung erstellen  -> v");
+                System.out.println("Film löschen           -> fl");
+                System.out.println("Vorstellung löschen    -> vl");
+                System.out.println("Zurück ins Hauptmenü   -> back");
 
-            switch (input.toLowerCase()){
-                case "f":
-                    admin.addFilm();
-                case "v":
-                    admin.addVorstellung();
-                case "fl":
-                    admin.deleteFilm();
-                case "vl":
-                    admin.deleteVorstellung();
-                case "exit":
-                    System.out.println("zurück zum");
-                default:
-                    System.out.println("Ungültige Eingabe");
-            }
+                input = scan.nextLine();
 
-        }while (!input.equalsIgnoreCase("exit"));
+                switch (input.toLowerCase()){
+                    case "f":
+                        admin.addFilm();
+                        break;
+                    case "v":
+                        admin.addVorstellung();
+                        break;
+                    case "fl":
+                        admin.deleteFilm();
+                        break;
+                    case "vl":
+                        admin.deleteVorstellung();
+                        break;
+                    case "back":
+                        System.out.println("zurück zum");
+                        break;
+                    default:
+                        System.out.println("Ungültige Eingabe");
+                }
+
+            }while (!input.equalsIgnoreCase("back"));
+        }
+
     }
 
     public void userMenu(){
